frontend: add `X-Frame-Options` HTTP response header
Credit to SHIVAM KHAMBE for demonstrating a clickjacking
vulnerability related to permissive framing options.

We had considered enabling this previously during review
of findings from the Mozilla HTTP Observatory, but deemed
it low risk.  Receiving a PoC has changed our mind.

We are aware that a `frame-ancestors` directive is
available in CSP2 -- and that neither that directive nor
X-Frame-Options are currently valid to place in HTML
`<meta>` tags.  We would prefer to encode all relevant web
security settings within the HTML/application bundle, so
that they are transparent to end-users and easy to archive.

Therefore we want to duplicate the setting in both the HTTP
response headers (less transparent, but effective) and also
in the HTML `<meta>` tag (ignored, per spec, but
transparent).

We anticipate some risk of conflict if we configure a CSP
policy in both the HTTP header and also the HTML -- and in
addition, there is a complexity that our HTML CSP is
dynamically generated, in order to support SubResource
Integrity (SRI) -- scripting a way to write those dynamic
generated elements into our `haproxy.cfg` during redeploys
of the frontend web application isn't something we want to
implement at the moment.

As a result, we are using the legacy/older X-Frame-Options
HTTP header instead of the CSP2 frame-ancestors directive.

Refs:
  - https://developer.mozilla.org/en-US/observatory
  - openculinary/frontend@d713ef146ae60d3002c10e9c105e5ebfe40e4ebc
  - CAEVaxk6DjdUTSpMSpMkiLoMmgd+=HmbB81BOOU_-e=iCqY0K5A (private)